requirements.txt
python-telegram-bot>=21.0
APScheduler>=3.10
python-dotenv>=1.0
